Fix to #6576 - Query: issue warning for queries with paging operations (First, Take, Skip) without OrderBy
Performing paging operations without OrderBy may lead to unpredictable results and could be hard to track down. Fix is to issue a warning in the logger if such situation happens.
Also added some refactoring to query model and expression printers - added option to remove formatting from the output string as well as limit the output to certain number of characters.